feat!: Add countdowns to competition/deadline on / instead of aerospacejam.org
For anyone curious as to why all links in the docs to `aerospacejam.org` were stripped out all of the sudden, it's actually because of some surprising bureaucracy at the WSGC that caught me off-guard - apparently we aren't allowed to have our own website and have to move back to the one with Carroll that you all are unfortunately already familiar with! We'll see if we can go back to it... but I don't have high hopes.